tags: avoid dependencies in base.TagDB
This makes it easilly importable anywhere.
- auto-completion code moved to autocomplete.py
- caching code more to storage.py

Another option could be to keep a light interface
and have a less abstract class with a few pre-filled methods.
